+++ scoped-gc/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+- **[Fix]** Fix lifetime elision error on `GcScope::alloc` ([#3](https://github.com/open-flash/rust-scoped-gc/pull/3))
+
 ## 0.1.2 (2018-11-26)
 
 - **[Fix]**: Add `Trace` impl for `BTreeMap`